backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
gspread-asyncio
python-dotenv
pydantic
anyio
python-multipart
aiofiles
python-magic-bin
Pillow
structlog
pydantic-settings
supabase